refactor: restructure org review layout - move risk flags above metrics
- Move account risk flags (Morocco, MAD) from detail view to overview section right column, displayed above payment metrics
- Extract dimension breakdown from organization review card, keep inline but render 1 per row instead of 2-column grid
- Remove recommended actions from dimension cards across all sections (overview, review, reviews history)
- Clean up unused imports in detail view
